More split cleanups.
git-svn-id: svn://svn.code.sf.net/p/bogofilter/code/trunk@3465 046ba1d7-8659-4f4f-8137-08d775d845e8

+++ bogofilter/src/buff.h
@@ -23,18 +23,18 @@ typedef struct {
     uint   size;	/* capacity */
 } buff_t;
 
-extern buff_t  *buff_new(byte *buff, size_t used, size_t size);
-extern buff_t  *buff_init(buff_t *self, byte *buff, size_t used, size_t size);
+extern buff_t  *buff_new(byte *buff, uint used, uint size);
+extern buff_t  *buff_init(buff_t *self, byte *buff, uint used, uint size);
 extern void 	buff_free(buff_t *self);
 extern void	buff_free_text(buff_t *self);
 extern int	buff_add(buff_t *self, word_t *in);
 extern buff_t  *buff_dup(const buff_t *self);
 extern int 	buff_cmp(const buff_t *t1, const buff_t *t2);
 extern int	buff_fgetsl(buff_t *self, FILE *);
-extern int	buff_fgetsln(buff_t *self, FILE *, size_t);
-extern void 	buff_puts(const buff_t *self, size_t width, FILE *fp);
+extern int	buff_fgetsln(buff_t *self, FILE *, uint);
+extern void 	buff_puts(const buff_t *self, uint width, FILE *fp);
 
-extern void 	buff_shift(buff_t *self, byte *start, size_t length);
+extern void 	buff_shift(buff_t *self, byte *start, uint length);
 
 #endif	/* BUFF_H */
